+++ extensions/tailwindcss/src/search-classes.tsx
@@ -1,39 +1,33 @@
 import { Action, ActionPanel, List } from "@raycast/api";
-import classes from "./classes";
+import output from "../generator/output.json";
+import { useState } from "react";
+
+const classes = Object.entries(output).map(([selector, classes]) => ({
+  selector,
+  classes,
+}));
 
 export default function SearchClasses() {
+  const [query, setQuery] = useState("");
+  const results = classes.filter((c) => c.selector.includes(query)).slice(0, 100);
   return (
-    <List>
-      {Object.entries(classes).map(([title, item]) => (
-        <List.Section key={title} title={title}>
-          {Object.entries(item).map(([className, style]) => (
-            <List.Item
-              title={className}
-              key={className}
-              subtitle={style.description}
-              accessories={[{ text: style.value }]}
-              actions={
-                <ActionPanel>
-                  <Action.CopyToClipboard
-                    title="Copy Class Name"
-                    content={className}
-                    shortcut={{ modifiers: ["cmd"], key: "." }}
-                  />
-                  <Action.CopyToClipboard
-                    title="Copy Description"
-                    content={style.description}
-                    shortcut={{ modifiers: ["cmd", "shift"], key: "." }}
-                  />
-                  <Action.CopyToClipboard
-                    title="Copy Generated CSS"
-                    content={style.value}
-                    shortcut={{ modifiers: ["cmd"], key: "," }}
-                  />
-                </ActionPanel>
-              }
-            />
-          ))}
-        </List.Section>
+    <List searchText={query} onSearchTextChange={setQuery} searchBarPlaceholder="Search all classes...">
+      {results.map(({ selector, classes }) => (
+        <List.Item
+          title={selector}
+          key={selector}
+          accessories={[{ text: classes }]}
+          actions={
+            <ActionPanel>
+              <Action.CopyToClipboard
+                title="Copy Selector"
+                content={selector}
+                shortcut={{ modifiers: ["cmd"], key: "." }}
+              />
+              <Action.CopyToClipboard title="Copy CSS" content={classes} shortcut={{ modifiers: ["cmd"], key: "," }} />
+            </ActionPanel>
+          }
+        />
       ))}
     </List>
   );