+++ extensions/tailwindcss/changelog.md
@@ -1,7 +1,9 @@
 # Tailwind CSS Changelog
 
-## [Fixed typo] - 2022-09-12
+## [Generate all CSS classes] - 2022-11-27
+- Generate all CSS classes for the `Search Classes` command
 
+## [Fixed typo] - 2022-09-12
 - Renamed grid-col- to grid-cols
  
 ## [Search Colors and Components command] - 2022-06-27